:recycle: Rename to diaries
